better handling of multiline logs
Before this commit lines with no date were printed if the last line with
a date was still in range. But when using --sort order --interleave this
would lead to strange sorting. dategrep now reads a multiline log as
single entry and prints it together. Had to refactor Peekable to handle
that.